Fixed some Android Studio warnings.

+++ app/src/main/cpp/MobileRT/Cameras/PerspectiveCamera.cpp
@@ -11,8 +11,8 @@ using namespace MobileRT;
 PerspectiveCamera::PerspectiveCamera(const Point3D& position, const float hFov, const float vFov) :
         position_(position),
         // convert to radians
-        hFov_(hFov * PI / 180.0f),
-        vFov_(vFov * PI / 180.0f)
+        hFov_(static_cast<float>(hFov * PI / 180.0f)),
+        vFov_(static_cast<float>(vFov * PI / 180.0f))
 {
 }
